requirements.txt
rdflib>=6.0.0
pandas>=1.5.0
openai>=1.0.0
python-dotenv>=1.0.0
pyvis>=0.3.2
